+++ doczrc.js
@@ -4,7 +4,7 @@ export default {
   plugins: [
     css({
       preprocessor: 'postcss',
-      cssmodules: true
-    })
-  ]
+      cssmodules: true,
+    }),
+  ],
 };