+++ src/__mocks__/utils.js
@@ -1,0 +1,6 @@
+const actualUtils = require.requireActual('../utils');
+
+module.exports = Object.assign(actualUtils, {
+  // eslint-disable-next-line no-undef
+  getElementHeight: jest.fn(),
+});